+++ src/main/kotlin/com/getkeepsafe/dexcount/OutputFormat.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 KeepSafe Software
+ * Copyright (C) 2015-2016 KeepSafe Software
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,17 +16,27 @@
 
 package com.getkeepsafe.dexcount
 
-class IOUtil {
-    static def drainToFile(InputStream stream, File file) {
-        stream.withStream { input ->
-            file.withOutputStream { output ->
-                def buf = new byte[4096]
-                def read
-                while ((read = input.read(buf)) != -1) {
-                    output.write(buf, 0, read)
-                }
-                output.flush()
-            }
-        }
-    }
+/**
+ * Specifies what format the task output should take.
+ */
+enum class OutputFormat(val extension: String) {
+    /**
+     * Specifies that method counts will be printed in a flat list of packages.
+     */
+    LIST(".txt"),
+
+    /**
+     * Specifies that the output will be pretty-printed as an tree.
+     */
+    TREE(".txt"),
+
+    /**
+     * Specifies that the output will be a pretty-printed JSON object.
+     */
+    JSON(".json"),
+
+    /**
+     * Specifies that output will be a YAML document.
+     */
+    YAML(".yml")
 }